remove bfloat

+++ train.py
@@ -151,7 +151,7 @@ def train(epoch):
         inputs, targets = inputs.to(device), targets.to(device)
         optimizer.zero_grad()
         
-        with torch.autocast(device_type=device, dtype=torch.bfloat16):
+        with torch.autocast(device_type=device):
             outputs = net(inputs)
 
             # We use L2 regularization instead of weight decay
